requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
google-genai==0.3.0
pandas
plotly